+++ selfie-runner-junit5/src/main/kotlin/com/diffplug/selfie/junit5/WriteTracker.kt
@@ -22,6 +22,7 @@ import java.nio.file.Files
 import java.nio.file.Path
 import java.util.stream.Collectors
 import kotlin.io.path.name
+import org.opentest4j.AssertionFailedError
 
 /** Represents the line at which user code called into Selfie. */
 data class CallLocation(val clazz: String, val method: String, val file: String?, val line: Int) :
@@ -101,8 +102,10 @@ internal class InlineWriteTracker : WriteTracker<CallLocation, LiteralValue<*>>(
   fun record(call: CallStack, literalValue: LiteralValue<*>, layout: SnapshotFileLayout) {
     recordInternal(call.location, literalValue, call, layout)
     if (literalValue.expected != null) {
-      throw UnsupportedOperationException(
-          "`.toBe() didn't match! Change to `toBe_TODO()` to record a new value until https://github.com/diffplug/selfie/pull/49 is merged.")
+      throw AssertionFailedError(
+          "`.toBe() didn't match! Change to `toBe_TODO()` to record a new value until https://github.com/diffplug/selfie/pull/49 is merged",
+          literalValue.expected,
+          literalValue.actual)
     }
   }
   fun hasWrites(): Boolean = writes.isNotEmpty()