+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Fixed
+* When `toBe()` snapshots fail in write mode, they now include the expected / actual values, to fill the gap until #49 is finished.
 
 ## [0.2.0] - 2023-12-27
 ### Added